#7404 - Refactoring
