+++ sormas-ui/src/main/java/de/symeda/sormas/ui/vaccination/list/event/VaccinationListChangedEventListener.java
@@ -1,4 +1,4 @@
-package de.symeda.sormas.ui.vaccination.list;
+package de.symeda.sormas.ui.vaccination.list.event;
 
 import java.lang.reflect.Method;
